Extract reusable scoped helpers for server tests
