+++ obs-studio-server/tests/test-osn-encoder-settings.cpp
@@ -9,47 +9,9 @@
 #include "obs-setup.hpp"
 #include "osn-error.hpp"
 #include "osn-video-encoder.hpp"
+#include "scoped-helpers.hpp"
 
 namespace {
-
-class ScopedConfigValue {
-public:
-	ScopedConfigValue(config_t *config, const char *section, const char *name, const char *value)
-		: config(config), section(section), name(name), hadUserValue(config_has_user_value(config, section, name))
-	{
-		const char *previous = config_get_string(config, section, name);
-		previousValue = previous ? previous : "";
-		config_set_string(config, section, name, value);
-	}
-
-	~ScopedConfigValue()
-	{
-		if (hadUserValue)
-			config_set_string(config, section, name, previousValue.c_str());
-		else
-			config_remove_value(config, section, name);
-	}
-
-private:
-	config_t *config;
-	const char *section;
-	const char *name;
-	bool hadUserValue;
-	std::string previousValue;
-};
-
-class ScopedEncoder {
-public:
-	explicit ScopedEncoder(obs_encoder_t *encoder) : encoder(encoder) {}
-	~ScopedEncoder()
-	{
-		osn::VideoEncoder::Manager::GetInstance().free(encoder);
-		obs_encoder_release(encoder);
-	}
-
-private:
-	obs_encoder_t *encoder;
-};
 
 std::string readSimpleStreamingSettings()
 {
@@ -67,11 +29,11 @@ TEST_CASE("Encoder settings preserve native defaults during creation and replace
 {
 	osn::tests::ObsSetup setupOBS;
 	config_t *config = ConfigManager::getInstance().getBasic();
-	ScopedConfigValue mode(config, "Output", "Mode", "Simple");
-	ScopedConfigValue selectedEncoder(config, "SimpleOutput", "StreamEncoder", "x264");
-	ScopedConfigValue advanced(config, "SimpleOutput", "UseAdvanced", "false");
-	ScopedConfigValue preset(config, "SimpleOutput", "Preset", "fast");
-	ScopedConfigValue customSettings(config, "SimpleOutput", "x264Settings", "scenecut=0");
+	osn::tests::ScopedConfigValue mode(config, "Output", "Mode", "Simple");
+	osn::tests::ScopedConfigValue selectedEncoder(config, "SimpleOutput", "StreamEncoder", "x264");
+	osn::tests::ScopedConfigValue advanced(config, "SimpleOutput", "UseAdvanced", "false");
+	osn::tests::ScopedConfigValue preset(config, "SimpleOutput", "Preset", "fast");
+	osn::tests::ScopedConfigValue customSettings(config, "SimpleOutput", "x264Settings", "scenecut=0");
 
 	for (bool explicitPreset : {false, true}) {
 		INFO("UseAdvanced = " << explicitPreset);
@@ -90,7 +52,7 @@ TEST_CASE("Encoder settings preserve native defaults during creation and replace
 		REQUIRE((ErrorCode)response[0].value_union.ui64 == ErrorCode::Ok);
 		const uint64_t encoderId = response[1].value_union.ui64;
 		obs_encoder_t *encoder = osn::VideoEncoder::Manager::GetInstance().find(encoderId);
-		ScopedEncoder encoderCleanup(encoder);
+		osn::tests::ScopedEncoder encoderCleanup(encoder);
 		REQUIRE(encoder);
 		OBSDataAutoRelease settings = obs_encoder_get_settings(encoder);
 		REQUIRE(settings);
